refactor io methods
